pyproject.toml
[project]
name = "pandasai-playaround"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandasai>=3.0.0",
    "pandasai-litellm>=0.0.1",
    "pandasai-openai>=0.1.6",
    "python-dotenv>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
